+++ dflib-echarts/src/main/java/org/dflib/echarts/TooltipTrigger.java
@@ -1,0 +1,9 @@
+package org.dflib.echarts;
+
+/**
+ * @since 1.0.0-M22
+ */
+public enum TooltipTrigger {
+
+    item, axis, none
+}