+++ dflib-echarts/src/main/java/org/dflib/echarts/render/option/tooltip/TooltipAxisPointerModel.java
@@ -1,0 +1,17 @@
+package org.dflib.echarts.render.option.tooltip;
+
+/**
+ * @since 1.0.0-M22
+ */
+public class TooltipAxisPointerModel {
+
+    private final String type;
+
+    public TooltipAxisPointerModel(String type) {
+        this.type = type;
+    }
+
+    public String getType() {
+        return type;
+    }
+}